Add RpnSolver class for Reverse Polish Notation expressions
The RpnSolver class was added to handle evaluation of Reverse Polish Notation expressions within our StringSolver project. This class includes a mechanism to store operations in an easily extensible way, and has a main Solve method that leverages a stack for solving RPN expressions. This addition is part of a larger effort to support various string solving strategies in our library.