Add formula processing functionality
Implemented a new Input Classifier to check if the given input format can be treated as a formula by using Regex. This greatly improves performance as Regex is now converted at compile time using source generator. Also, added new handling methods in the main program to manage formula, variable assignments and calculate results for operations. The new implementation will replace and evaluate any instances of the variable in the formula. For invalid input or unknown variables, appropriate error messages will be displayed to the user.